misc(cli): refactor functions into utils and config dirs, merge yargs options
Moved filed to respective config or utils dir, cleaned up cli.js and removed yargs options and
merged in a single file

ISSUES CLOSED: #770
