+++ bin/config/webpackConfigurationSchema.json
