+++ bin/utils/prepareOptions.js
