+++ bin/utils/prompt-command.js
